requirements.txt
blinker==1.6.2
blurhash==1.1.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
Flask==2.3.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Mastodon.py==1.8.1
packaging==23.1
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.31.0
schedule==1.2.0
six==1.16.0
urllib3==2.0.4
Werkzeug==2.3.6
